+++ src/util/errors.js
@@ -15,6 +15,7 @@ import {
   ERROR_CODE_TRANSACTION_ALREADY_REVIEWED_BY_CUSTOMER,
   ERROR_CODE_TRANSACTION_ALREADY_REVIEWED_BY_PROVIDER,
   ERROR_CODE_PAYMENT_FAILED,
+  ERROR_CODE_CHARGE_ZERO_PAYIN,
   ERROR_CODE_EMAIL_TAKEN,
   ERROR_CODE_EMAIL_NOT_FOUND,
   ERROR_CODE_EMAIL_NOT_VERIFIED,
@@ -112,11 +113,20 @@ export const isTransactionInitiateBookingTimeNotAvailableError = error =>
   hasErrorWithCode(error, ERROR_CODE_TRANSACTION_BOOKING_TIME_NOT_AVAILABLE);
 
 /**
+ * Check if the given API error (from `sdk.transaction.initiate()` or
+ * `sdk.transaction.initiateSpeculative()`) is due to payment being zero.
+ */
+export const isTransactionZeroPaymentError = error =>
+  hasErrorWithCode(error, ERROR_CODE_CHARGE_ZERO_PAYIN);
+
+/**
  * Check if the given API error (from `sdk.transaction.initiate()`) is
  * due to the transaction total amount being too low for Stripe.
  */
 export const isTransactionInitiateAmountTooLowError = error => {
-  return responseAPIErrors(error).some(apiError => {
+  const isZeroPayment = isTransactionZeroPaymentError(error);
+
+  const tooLowAmount = errorAPIErrors(error).some(apiError => {
     const isPaymentFailedError =
       apiError.status === 402 && apiError.code === ERROR_CODE_PAYMENT_FAILED;
     let isAmountTooLow = false;
@@ -134,6 +144,26 @@ export const isTransactionInitiateAmountTooLowError = error => {
 
     return isPaymentFailedError && isAmountTooLow;
   });
+
+  return isZeroPayment || tooLowAmount;
+};
+
+/**
+ * Check if the given API error (from `sdk.transaction.initiate()`) is
+ * due to other error in Stripe.
+ */
+export const transactionInitiateOrderStripeErrors = error => {
+  if (error) {
+    return errorAPIErrors(error).reduce((messages, apiError) => {
+      const isPaymentFailedError =
+        apiError.status === 402 && apiError.code === ERROR_CODE_PAYMENT_FAILED;
+      const hasStripeError = apiError && apiError.meta && apiError.meta.stripeMessage;
+      const stripeMessageMaybe =
+        isPaymentFailedError && hasStripeError ? [apiError.meta.stripeMessage] : [];
+      return [...messages, ...stripeMessageMaybe];
+    }, []);
+  }
+  return null;
 };
 
 /**